# backend/openapi/generated-clients/model/update-task-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Taskbook API
 * API for a sample project which randomly mixes Kotlin and Java classes to show the interoperability capabilities. Also includes openapi client generation for a Vue3 UI
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { TaskState } from './task-state';

/**
 * 
 * @export
 * @interface UpdateTaskDto
 */
export interface UpdateTaskDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateTaskDto
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof UpdateTaskDto
     */
    'description': string;
    /**
     * 
     * @type {TaskState}
     * @memberof UpdateTaskDto
     */
    'state': TaskState;
}