# backend/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Taskbook API",
    "description": "API for a sample project which randomly mixes Kotlin and Java classes to show the interoperability capabilities. Also includes openapi client generation for a Vue3 UI",
    "contact": {
      "name": "Mewgrammer",
      "url": "https://github.com/Mewgrammer"
    },
    "license": {
      "name": "MIT Licence",
      "url": "https://github.com/thombergs/code-examples/blob/master/LICENSE"
    },
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "tags": [
    {
      "name": "task",
      "description": "API to perform CRUD Operations on Tasks"
    }
  ],
  "paths": {
    "/api/tasks/{id}": {
      "put": {
        "tags": [
          "task"
        ],
        "summary": "Updates a Task",
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Task not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid update-data supplied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "200": {
            "description": "The Task was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "task"
        ],
        "summary": "Deletes a Task with a given ID",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Task not found"
          },
          "200": {
            "description": "The Task was deleted"
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "tags": [
          "task"
        ],
        "summary": "Gets all Tasks",
        "operationId": "getTasks",
        "responses": {
          "200": {
            "description": "Returns all Tasks for the requesting user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          },
          "406": {
            "description": "Task already exists",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid data supplied",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "task"
        ],
        "summary": "Update a Task",
        "operationId": "addTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "406": {
            "description": "Task already exists",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data supplied",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "200": {
            "description": "The Task was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TaskState": {
        "type": "string",
        "enum": [
          "PENDING",
          "IN_PROGRESS",
          "FINISHED"
        ]
      },
      "UpdateTaskDto": {
        "required": [
          "description",
          "state",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TaskState"
          }
        }
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskState"
          },
          "creator": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTaskDto": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskState"
          }
        }
      }
    },
    "securitySchemes": {
      "api": {
        "type": "http",
        "in": "header",
        "scheme": "basic"
      }
    }
  }
}